Add task popup
Add task popup now functional. Next step is to make view task/edit task functional.

+++ Services/TaskService.cs
@@ -62,8 +62,11 @@ namespace BlazorTaskManager.Services
         {
             using (var context = _dbContextFactory.CreateDbContext())
             {
-                context.TaskItem.Remove(task);
-                context.SaveChanges();
+                if(context.TaskItem.Contains(task))
+                {
+                    context.TaskItem.Remove(task);
+                    context.SaveChanges();
+                }
             }
         }
